RavenDB-7322 - moved longest tests from Fast to Slow, - AttachmentsHugeFiles tests will only start in our nightly builds
