+++ test/StressTests/Client/Attachments/AttachmentsHugeFiles.cs
@@ -3,13 +3,14 @@ using System.IO;
 using FastTests;
 using FastTests.Client.Attachments;
 using Microsoft.AspNetCore.Http.Features;
+using Tests.Infrastructure;
 using Xunit;
 
 namespace StressTests.Client.Attachments
 {
     public class AttachmentsHugeFiles : NoDisposalNeeded
     {
-        [Theory]
+        [NightlyBuildTheory]
         [InlineData(FormOptions.DefaultMultipartBodyLengthLimit * 2, "vEbE0Uh02lIPx/cEFBagkmepLTP0nWWYX5+exkt9yoE=")] // 256 MB
         [InlineData(2.5 * 1024 * 1024 * 1024, "2ssXqJM7lbdDpDNkc2GsfDbmcQ6CXdgP6/LFmLtFCT4=")] // 2.5 GB
         public void BatchRequestWithLongMultiPartSections(long size, string hash)
@@ -29,7 +30,7 @@ namespace StressTests.Client.Attachments
             }
         }
 
-        [Theory]
+        [NightlyBuildTheory]
         [InlineData(2.5 * 1024 * 1024 * 1024, "2ssXqJM7lbdDpDNkc2GsfDbmcQ6CXdgP6/LFmLtFCT4=")] // 2.5 GB
         public void SupportHugeAttachment(long size, string hash)
         {
